kotlin: Merge *ListOptions into resource modules
Easier to replace with generated files this way, as we'll have one file per resource like in other libs.
