--- kotlin/lib/src/main/kotlin/EndpointStatsOptions.kt
@@ -1,8 +1,0 @@
-package com.svix.kotlin
-
-import java.time.OffsetDateTime
-
-class EndpointStatsOptions {
-    var since: OffsetDateTime? = null
-    var until: OffsetDateTime? = null
-}
