+++ kotlin/lib/src/main/kotlin/Message.kt
@@ -6,6 +6,32 @@ import com.svix.kotlin.models.ApplicationIn
 import com.svix.kotlin.models.ListResponseMessageOut
 import com.svix.kotlin.models.MessageIn
 import com.svix.kotlin.models.MessageOut
+import java.time.OffsetDateTime
+
+class MessageListOptions : ListOptions() {
+    var eventTypes: List<String>? = null
+    var before: OffsetDateTime? = null
+    var after: OffsetDateTime? = null
+    var channel: String? = null
+    var withContent: Boolean? = null
+    var tag: String? = null
+
+    fun eventTypes(eventTypes: List<String>) = apply { this.eventTypes = eventTypes }
+
+    fun before(before: OffsetDateTime) = apply { this.before = before }
+
+    fun after(after: OffsetDateTime) = apply { this.after = after }
+
+    fun channel(channel: String) = apply { this.channel = channel }
+
+    override fun iterator(iterator: String) = apply { super.iterator(iterator) }
+
+    override fun limit(limit: Int) = apply { super.limit(limit) }
+
+    fun withContent(withContent: Boolean) = apply { this.withContent = withContent }
+
+    fun tag(tag: String) = apply { this.tag = tag }
+}
 
 class Message internal constructor(token: String, options: SvixOptions) {
     val api = MessageApi(options.serverUrl)
@@ -68,7 +94,7 @@ class Message internal constructor(token: String, options: SvixOptions) {
 }
 
 /**
- * Creates a [MessageIn] with a raw string payload.
+ * Creates a [MessageIn] with a pre-serialized payload.
  *
  * The payload is not normalized on the server. Normally, payloads are required to be JSON, and Svix
  * will minify the payload before sending the webhooks (for example, by removing extraneous
