--- kotlin/lib/src/main/kotlin/OperationalWebhookEndpointListOptions.kt
@@ -1,13 +1,0 @@
-package com.svix.kotlin
-
-import com.svix.kotlin.models.Ordering
-
-class OperationalWebhookEndpointListOptions : ListOptions() {
-    var order: Ordering? = null
-
-    fun order(order: Ordering) = apply { this.order = order }
-
-    override fun iterator(iterator: String) = apply { super.iterator(iterator) }
-
-    override fun limit(limit: Int) = apply { super.limit(limit) }
-}
